+++ test/Faker/Provider/ComputerComponentTest.php
@@ -5,7 +5,7 @@ namespace Faker\Test\Provider;
 use Faker\Provider\ComputerComponents;
 use PHPUnit\Framework\TestCase;
 
-final class ComputerComponentsTest extends TestCase
+final class ComputerComponentTest extends TestCase
 {
 	public function testRandomComponent()
 	{
